requirements.txt
asgiref==3.6.0
black==23.1.0
click==8.1.3
Django==4.0.4
djangorestframework==3.14.0
isort==5.12.0
mypy-extensions==0.4.3
packaging==23.0
pathspec==0.11.0
platformdirs==2.6.2
pytz==2022.7.1
sqlparse==0.4.3
